size

+++ hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 char charType;
 int intType;
 long int longintType;
-long long longlongType;
+long double longlongType;
 float floatType;
 printf("Size of char: %zu byte\n", sizeof(charType));
 printf("Size of int: %zu bytes\n", sizeof(intType));
